bug fix
fixed the bug in the movie edit / scrape fanart. Not saving fanart and
poster on manual edit
added wait cursor on image download after selection in manual edit
fixed error on + for trailers. There was a conflicting (old) variable.
changed name of PostScraperCapabilities in ScraperCapabilites as is not
correct anymore to bind it to Poster only