changed select * with select <fields>
The * will return the fields in the creation order. This creates issues
when the columns are added at runtime (i.e. DB updated)
Now all queries are explicit so the column orders are strict.
If the DB is updated then the queries need to be updated if the column
is requested