removed: RefreshAllMovieSets after movie scraping
The original idea was to look for already existing (not deleted) images
for moviesets. Unfortunately, so we lost the "New" tag of all new
moviesets. Since it is possible to remove the moviesets images with
moviesets, most user will no more have old/undeleted movieset images,
Therefore, it is now no longer look for old images after movie scraping.