code: changed all UInt64 and Int32 to Int64 for SQLite handling
UInt32/64 is not existing in SQLite, so it makes no sense to use it as
database type. Also SQLite tries to get the best database type for each
value even what is set as database type for this column. Only use Int64
should be the best and safest way to handle all Integer values. Read
wiki for more inforamtions:
https://www.sqlite.org/datatype3.html#section_3_2